Refactor ONNX graph surgery base classes
Move the surgeon registry and shared proto/rewrite-rule base classes into a dedicated package so new transformations can be implemented in focused modules without growing graph_surgeries.py.
