+++ olive/passes/onnx/graph_surgery/__init__.py
@@ -1,0 +1,7 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+# --------------------------------------------------------------------------
+from olive.passes.onnx.graph_surgery.base import ProtoSurgeon, RewriteRuleSurgeon, Surgeon
+
+__all__ = ["ProtoSurgeon", "RewriteRuleSurgeon", "Surgeon"]
